wrote cython code for casting into a cube. Really fastgit status
